Refactor: scss 파일 이동
