+++ src/components/header/Header.module.scss
@@ -1,5 +1,6 @@
 .Header {
   margin-top: 1rem;
+  margin-bottom: 5rem;
   width: 100%;
 
   &__box {
@@ -51,6 +52,7 @@
 
   &__title {
     margin-top: 13rem;
+
     &__text {
       font-family: 'Heading-Now';
       font-size: 7rem;
