+++ src/components/layout/AppLayout.tsx
@@ -1,4 +1,4 @@
-import Style from '@/styles/layout/AppLayout.module.scss';
+import Style from '@/components/layout/AppLayout.module.scss';
 
 const AppLayout = (props: { children: React.ReactNode }) => {
   return (
